// frontend/src/types/chat.ts
export const MODERATION_CATEGORIES = [
  "OFFENSIVE",
  "OFF_BRAND", 
  "VIOLENCE",
  "NONE",
] as const;

// Derive the union type for ModerationCategory from the array
export type ModerationCategory = (typeof MODERATION_CATEGORIES)[number];
// ...
export interface GuardrailResultType {
  status: "IN_PROGRESS" | "DONE";
  testText?: string; 
  category?: ModerationCategory;
  rationale?: string;
}